updates to strain-detangler and luigi workflow
Removed list-writing as functions of split_reference and reduce, since this is more of a workflow task. Made it possible to customize reduce radius, band filter stringency, and the maximum number of strains to consider. Big luigi workflow wrapping up everything from mapping to strain deconvolution, which comes with a config file and a small test case.